% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttm_update.R
\name{dttm_update}
\alias{dttm_update}
\title{Merge Date and Time Columns into Datetime}
\usage{
dttm_update(x, date_col, time_col)
}
\arguments{
\item{x}{A data frame.}

\item{date_col}{Name of the date/datetime column (character).}

\item{time_col}{Name of the time column (character).}
}
\value{
The updated data frame.
}
\description{
Updates a date/datetime column using hour, minute and second values extracted
from another time column.
}
