% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_issue.R
\name{extract_issue}
\alias{extract_issue}
\title{Extract a Component from Safe/Quietly Results}
\usage{
extract_issue(x, component)
}
\arguments{
\item{x}{A named list where each element is itself a list containing
a common component to extract.}

\item{component}{Character scalar with the component name to extract.}
}
\value{
A data frame created by row-binding the selected component across
all list elements, with a \code{dataset} identifier column.
}
\description{
Extracts a specific component from a list of outputs usually created by
wrappers such as \code{purrr::safely()} or \code{purrr::quietly()}, binding results
across datasets.
}
