% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_column_types.R
\name{set_column_types}
\alias{set_column_types}
\title{Read Column Types Metadata}
\usage{
set_column_types(sheet = NULL)
}
\arguments{
\item{sheet}{Character scalar with the target spreadsheet name (for example,
\code{"Camera_trap"} or \code{"Underpasses"}).}
}
\value{
A character vector of column types.
}
\description{
Returns the predefined column type metadata stored internally in the package
for a given spreadsheet.
}
