% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo_matching.R
\name{stringdist_table}
\alias{stringdist_table}
\title{Build String Distance Table}
\usage{
stringdist_table(filenames_on_sheet, media)
}
\arguments{
\item{filenames_on_sheet}{Character vector of expected filenames.}

\item{media}{Character vector of media filenames.}
}
\value{
Tibble with distance matrix in wide format and \code{match_exactly}.
}
\description{
Computes Levenshtein distances between expected filenames and media
candidates.
}
