% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_column_types.R
\name{set_column_types}
\alias{set_column_types}
\title{Read Column Types Metadata}
\usage{
set_column_types(sheet = NULL)
}
\arguments{
\item{sheet}{Character scalar with the sheet name in
\code{support/column_types.xlsx}.}
}
\value{
A character vector of column types.
}
\description{
Reads column type definitions from \code{support/column_types.xlsx} for a given
sheet.
}
